Move FeedbackSelector out of ConversationMessageActions
